Move validation helpers to subfolder.
